[ :white_check_mark: ] | test: matrix - transformations > in sequence (math::matrix)
